+++ RedPandaIDE/compiler/stdincompiler.cpp
@@ -47,7 +47,8 @@ bool StdinCompiler::prepareForCompile()
     default:
         throw CompileError(tr("Can't find the compiler for file %1").arg(mFilename));
     }
-    mArguments += getLibraryArguments(fileType);
+    if (!mOnlyCheckSyntax)
+        mArguments += getLibraryArguments(fileType);
 
     if (!fileExists(mCompiler)) {
         throw CompileError(tr("The Compiler '%1' doesn't exists!").arg(mCompiler));